+++ src/private/main.cpp
@@ -1,9 +1,7 @@
-#include "game/headers/Game.h"
-#include "game/headers/GlobalVars.h"
+#include "game/Game.h"
+#include "game/GlobalVars.h"
 #include <filesystem>
 
-//temp include
-#include "utils/headers/TextureManager.h"
 int main()
 {
     Game::init(conf::window_size.x,conf::window_size.y, conf::worldSize.x,conf::worldSize.y);
