+++ src/private/utils/TextureManager.h
